+++ Source/EbsdLib/Utilities/ColorUtilities.cpp
@@ -126,7 +126,7 @@ EbsdLib::Rgb ColorUtilities::ConvertHSVtoRgb(float h, float s, float v)
     b = 0.0f;
   }
 
-  return EbsdLib::RgbColor::dRgb(r * 255, g * 255, b * 255, 0);
+  return EbsdLib::RgbColor::dRgb(static_cast<int>(r * 255), static_cast<int>(g * 255), static_cast<int>(b * 255), 0);
 }
 
 // -----------------------------------------------------------------------------
@@ -158,7 +158,7 @@ EbsdLib::Rgb ColorUtilities::Hsv2Rgb(float h, float s, float v)
     out.r = in.v;
     out.g = in.v;
     out.b = in.v;
-    return EbsdLib::RgbColor::dRgb(out.r * 255, out.g * 255, out.b * 255, 255);
+    return EbsdLib::RgbColor::dRgb(static_cast<int>(out.r * 255), static_cast<int>(out.g * 255), static_cast<int>(out.b * 255), 255);
   }
   hh = in.h;
   if(hh >= 360.0)
@@ -207,7 +207,7 @@ EbsdLib::Rgb ColorUtilities::Hsv2Rgb(float h, float s, float v)
     out.b = q;
     break;
   }
-  return EbsdLib::RgbColor::dRgb(out.r * 255, out.g * 255, out.b * 255, 255);
+  return EbsdLib::RgbColor::dRgb(static_cast<int>(out.r * 255), static_cast<int>(out.g * 255), static_cast<int>(out.b * 255), 255);
 }
 
 // -----------------------------------------------------------------------------
@@ -216,13 +216,13 @@ EbsdLib::Rgb ColorUtilities::Hsv2Rgb(float h, float s, float v)
 std::vector<EbsdLib::Rgb> ColorUtilities::GenerateColors(int count, int saturation, int value)
 {
   std::vector<EbsdLib::Rgb> colors(count);
-  float s = static_cast<float>(saturation) / 255.0;
-  float v = static_cast<float>(value) / 255.0;
-  float increment = 360.0 / count;
+  float s = static_cast<float>(saturation) / 255.0f;
+  float v = static_cast<float>(value) / 255.0f;
+  float increment = 360.0f / count;
   for(float i = 0; i < static_cast<float>(count); i = i + 1.0f)
   {
     EbsdLib::Rgb c = ColorUtilities::Hsv2Rgb(i * increment, s, v);
-    colors[i] = c;
+    colors[static_cast<size_t>(i)] = c;
   }
   return colors;
 }
